+++ src/entities/tag/api/queries/index.ts
@@ -1,0 +1,3 @@
+import useTagsQuery from "./useTagsQuery"
+
+export { useTagsQuery }
